+++ views/transportation/bus/bus-line.js
@@ -3,8 +3,7 @@
 import React from 'react'
 import {View, Text, StyleSheet, Platform} from 'react-native'
 import type {BusLineType, FancyBusTimeListType} from './types'
-import getScheduleForNow from './get-schedule-for-now'
-import getSetOfStopsForNow from './get-set-of-stops-for-now'
+import {getScheduleForNow, getSetOfStopsForNow} from './lib'
 import zip from 'lodash/zip'
 import head from 'lodash/head'
 import last from 'lodash/last'
