+++ views/transportation/bus/lib/index.js
@@ -1,0 +1,2 @@
+export {getScheduleForNow} from './get-schedule-for-now'
+export {getSetOfStopsForNow} from './get-set-of-stops-for-now'
